fix(input): make mouse coordinate conversion explicit
The platform mouse location uses double-valued utility coordinates while the input backend exposes integer points. Explicit conversion preserves Windows POINT semantics and prevents the CI -Werror narrowing failure.


Model: GPT 5.6-Terra X-High
(cherry picked from commit f6b252438b99df082d036ffecbddab4e3e78e955)
(cherry picked from commit fa5962b048622677faf03c17a744fcfa673cb639)

+++ src/input.cpp
@@ -123,7 +123,10 @@ namespace input {
     mouse_input::point_t position() const override {
       auto &input = const_cast<platf::input_t &>(_input);
       const auto position = platf::get_mouse_loc(input);
-      return {position.x, position.y};
+      return {
+        static_cast<int>(position.x),
+        static_cast<int>(position.y),
+      };
     }
 
     void move_relative(mouse_input::point_t delta) override {
